+++ aoxp-server/src/test/java/com/ao/model/character/movement/QuietMovementStrategyTest.java
@@ -32,8 +32,8 @@ public class QuietMovementStrategyTest {
 
 	@Test
 	public void testMove() {
-		final Position pos = new Position((byte) 50, (byte) 50, 1);
-		final Position target = new Position((byte) 60, (byte) 60, 1);
+		final Position pos = new Position((byte) 50, (byte) 50, (short) 1);
+		final Position target = new Position((byte) 60, (byte) 60, (short) 1);
 
 		movement.setTarget(target);
